+++ src/components/TabContainer/TabContainer.scss
@@ -1,5 +1,5 @@
 .tab-container {
   display: flex;
   overflow: hidden;
-  padding: 70px 0 37px 0;
+  padding-bottom: 37px;
 }
